==> script/1_get_unique_name.ipynb <==
import pandas as pd

eICU_med_path = '/home/qlunlp/ylh/database/eICU/medication.csv'
origin_med_pd = pd.read_csv(eICU_med_path, dtype={'drugname': str})


med_pd = pd.read_csv(eICU_med_path, usecols=['drugname'], dtype={'drugname': str})
med_pd.head()

len(med_pd)

med_pd.drop(med_pd[med_pd['drugname'].isnull()].index, inplace=True)
len(med_pd)

med_pd.drop_duplicates(subset=['drugname'], keep='first', inplace=True)
len(med_pd)

med_pd = med_pd.reset_index(drop=True)
len(med_pd)

med_pd.to_csv('/home/qlunlp/ylh/database/tmp/unique_drugname.csv', index=False)

drugname = pd.read_csv('/home/qlunlp/ylh/database/tmp/unique_drugname.csv', dtype={'drugname': str})
drugname.head()

for item in drugname['drugname']:
    cnt = len(origin_med_pd[origin_med_pd['drugname'] == item])
    if cnt == 0:
        print(f"Drugname {item} not found in original medication data.")    

## 映射文件

import pandas as pd
rxnorm_mappings_path = "/home/qlunlp/ylh/database/drug/rxnorm_mappings.txt"
origin_rxnorm_mappings = pd.read_csv(rxnorm_mappings_path, sep="|",usecols=['RXCUI','RXSTRING'], dtype={'RXSTRING': str, 'RXCUI': str})
origin_rxnorm_mappings.head()

rxnorm_mapping = origin_rxnorm_mappings[['RXCUI', 'RXSTRING']]
len(rxnorm_mapping)

rxnorm_mapping.dropna(subset=['RXCUI', 'RXSTRING'], inplace=True)
rxnorm_mapping.drop_duplicates(subset=['RXSTRING'],inplace=True)
len(rxnorm_mapping)

rxnorm_mapping.head()

import re
def normalize_text(s: str) -> str:
    """基础规范化：小写，去多余空格，标准化常见缩写等。"""
    s = s.lower()
    # 去掉多余连字符，两边留空格方便匹配
    s = s.replace('-', ' ')
    s = s.replace('/', ' / ')
    # 一些常见缩写展开
    s = s.replace(' ipratropium br ', ' ipratropium bromide ')
    s = s.replace(' tabs ', ' tablet ')
    s = s.replace(' tbec ', ' tablet enteric ')
    s = s.replace(' inj ', ' injection ')
    s = re.sub(r'\s+', ' ', s)
    return s.strip()

rxnorm_mapping['RXSTRING'] = rxnorm_mapping['RXSTRING'].astype(str).apply(normalize_text)
rxnorm_mapping.drop_duplicates(subset=['RXSTRING'], keep='first', inplace=True)
rxnorm_mapping = rxnorm_mapping.reset_index(drop=True)
rxnorm_mapping

len(rxnorm_mapping)

rxnorm_mapping.to_csv('/home/qlunlp/ylh/database/eICU/process/tmp/processed_rxnorm_mappings.csv', index=False)
