==> script/3_mappings.ipynb <==
import pandas as pd

ori_mappings_file = '/home/qlunlp/ylh/database/eICU/process/tmp/drug_rxcui_mapping.csv' # TODO: update
ndc2rxcui_file = '/home/qlunlp/ylh/database/drug/ndc2RXCUI.txt'
ndc2dbid_file = '/home/qlunlp/ylh/database/drug/ndc2db_all.csv'
drugbank_file = '/home/qlunlp/ylh/database/drug/drugbank_drugs_info.csv'

ori_mappings = pd.read_csv(ori_mappings_file, sep=',', dtype={'RXCUI': str})

with open(ndc2rxcui_file, 'r') as f:
    ndc2rxcui = eval(f.read())

rxcui2ndc = {v: k[0:-2] for k,v in ndc2rxcui.items()}
rxcui2ndc

ori_mappings["NDC"] = ori_mappings["RXCUI"].map(rxcui2ndc)
ori_mappings.dropna(inplace=True)
ori_mappings

ndc2dbid = pd.read_csv(ndc2dbid_file, sep=',', dtype={'NDC': str, 'drugbank_id': str})
ori_mappings = ori_mappings.merge(ndc2dbid, on='NDC', how='inner')
ori_mappings

drugbank_info = pd.read_csv(drugbank_file, sep=',', dtype=str, usecols=['drugbank_id', 'name', 'moldb_smiles'])
ori_mappings = ori_mappings.merge(drugbank_info, on='drugbank_id', how='inner')
ori_mappings.drop_duplicates(inplace=True)
ori_mappings.reset_index(drop=True, inplace=True)
ori_mappings

ori_mappings.to_csv('/home/qlunlp/ylh/database/eICU/process/tmp/all_mappings.csv', sep='\t', index=False)
