==> script/4_med_mappings.ipynb <==
import pandas as pd

df = pd.read_csv("/home/qlunlp/ylh/database/eICU/process/tmp/all_mappings.csv", sep="\t")  # 自动识别表头

# 改Drug Name为drug name
df = df.rename(columns={"Drug Name": "drugname"})
df.head()

medications_file = "/home/qlunlp/ylh/database/eICU/治疗/medication.csv"
med_df = pd.read_csv(medications_file, dtype={'drugname': str})
med_df.head()

med_df_atc = med_df.merge(df, on='drugname', how='inner')
med_df_atc.head()

len(med_df_atc)

med_df_atc = med_df_atc[med_df_atc['RXCUI'] != 'NO_MATCH']
len(med_df_atc)

RXCUI2atc4_file = '/home/qlunlp/ylh/database/drug/RXCUI2atc4.csv'

RXCUI2atc4 = pd.read_csv(RXCUI2atc4_file, sep=',', usecols=["RXCUI", "ATC4"])
RXCUI2atc4.drop_duplicates(subset=["RXCUI"], inplace=True)
med_df_atc.drop(index=med_df_atc[med_df_atc["RXCUI"].isin([""])].index, axis=0, inplace=True)


med_df_atc["RXCUI"] = med_df_atc["RXCUI"].astype("int64")
med_df_atc = med_df_atc.reset_index(drop=True)
med_df_atc = med_df_atc.merge(RXCUI2atc4, on=["RXCUI"], how='inner')

len(med_df_atc)

med_df_atc["ATC4"] = med_df_atc["ATC4"].map(lambda x: x[:4])
med_df_atc = med_df_atc.rename(columns={"ATC4": "ATC3"})
med_df_atc = med_df_atc.drop_duplicates()
med_df_atc = med_df_atc.reset_index(drop=True)

len(med_df_atc)

med_df_atc.to_csv("/home/qlunlp/ylh/database/eICU/output/medication_mappings.csv", index=False)
